Refactor: Separate ECS handlers into resource-specific files
- Consolidate task definition handlers and types in task_definition.go
- Move task definition implementation from old mock handlers to storage-backed handlers
- Move cluster HTTP handlers from ecs_handler.go to cluster_handler.go
- Create task_handler.go for task-related operations
- Create service_handler.go for service-related operations
- Keep ecs_handler.go as the main routing file only
- Remove duplicate handler implementations
- This improves code organization and prevents ecs_handler.go from becoming too large
